requirements.txt
python-telegram-bot==20.7
mysql-connector-python==8.2.0
aiohttp==3.9.1
aiofiles==23.2.1
openpyxl==3.1.2
pandas==2.1.4
python-dotenv==1.0.0
opencv-python==4.8.1.78
moviepy==1.0.3 